requirements.txt
arrow==0.4.4
authomatic==0.0.8
Flask==0.10
flask-jsonpify==1.5.0
google-api-python-client==1.2
GoogleAppEngineMapReduce==1.9.22.0
isoweek==1.3.0
pytz==2014.7
